refactor(domain): update models based on review feedback
Addresses comments from the pull request review:

- Renames `time_unit.rs` to `duration.rs` for clarity.
- Refactors the `Duration` model to simplify its API, renaming fields and methods and making `TimeUnit` a private enum.
- Updates the `Step` model to use `u8` for the `position` field.
- Makes the `note` field mandatory in the `DurationRangeParameter` model.
- Removes unnecessary `pub use` statements from `models.rs`.

+++ backend/src/domain/models/step.rs
@@ -28,13 +28,13 @@ pub struct Step {
     id: StepId,
     trial_id: TrialId,
     name: Option<String>,
-    position: u32,
+    position: u8,
     started_at: Option<DateTime<Utc>>,
 }
 
 impl Step {
     /// 新しい Step を作成する
-    pub fn new(trial_id: TrialId, position: u32) -> Self {
+    pub fn new(trial_id: TrialId, position: u8) -> Self {
         Self {
             id: StepId::new(),
             trial_id,
@@ -49,7 +49,7 @@ impl Step {
         id: StepId,
         trial_id: TrialId,
         name: Option<String>,
-        position: u32,
+        position: u8,
         started_at: Option<DateTime<Utc>>,
     ) -> Self {
         Self {
@@ -74,7 +74,7 @@ impl Step {
         self.name.as_deref()
     }
 
-    pub fn position(&self) -> u32 {
+    pub fn position(&self) -> u8 {
         self.position
     }
